+++ packages/commerce-sdk-react/src/auth/index.ts
@@ -512,7 +512,7 @@ class Auth {
             this.set('customer_type', isGuest ? 'guest' : 'registered')
 
             accessToken = sfraAuthToken
-            // SFRA -> PWA access token cookie handoff is succesful so we clear the SFRA made cookies.
+            // SFRA -> PWA access token cookie handoff is successful so we clear the SFRA made cookies.
             // We don't want these cookies to persist and continue overriding what is in local store.
             this.clearSFRAAuthToken()
         }
@@ -591,6 +591,8 @@ class Auth {
      * store the data in storage.
      */
     private handleTokenResponse(res: TokenResponse, isGuest: boolean) {
+        // Delete the SFRA auth token cookie if it exists
+        this.clearSFRAAuthToken()
         this.set('access_token', res.access_token)
         this.set('customer_id', res.customer_id)
         this.set('enc_user_id', res.enc_user_id)